Update README.md
We decided on 45 days rather than 30 in order to maintain a little more history.  We did have a situation where a volunteer wanted to reach back to someone who called on his shift.  Fortunately we could retrieve the message, but we got lucky.  This will give us a bit more luck.